BirthdayCalendar refactoring stateful homepage widget to another class and creating separate package for it
